Refactor favorite module to use user ID instead of email, update favorite model and service, and improve pagination logic
- Updated favorite controller to use user ID for fetching and managing favorite books.
- Refactored favorite interface and model to replace customerEmail with userId.
- Removed unnecessary methods and optimized favorite service for better performance.
- Adjusted favorite routes to reflect changes in controller methods.
- Added health check route for application status monitoring.
- Enhanced order module with new endpoints for fetching, updating, and deleting orders.
- Improved user module to include user profile retrieval and role management.
- Removed review module as it was deemed unnecessary.
- Updated validation schemas across modules for consistency and clarity.

+++ src/modules/book/service/book.ts
@@ -11,8 +11,8 @@ import {
 } from "@/book/validation/book.js";
 import Category from "@/category/model/category.js";
 import Order from "@/order/model/order.js";
-import type { TUserRole } from "@/user/interface/user.js";
-import { User } from "@/user/model/user.js";
+import type { TUserRole } from "@/user/validation/user.js";
+import  User  from "@/user/model/user.js";
 import { UserRole } from "@/user/validation/user.js";
 import { getPaginatedData } from "@/utils/getPaginatedData.js";
 import {
@@ -106,7 +106,7 @@ const getBooks = async (queryPayload: unknown) => {
     if (user) {
       const { role, _id } = user;
 
-      if (roles.includes(role as Exclude<TUserRole, "user">)) {
+      if (roles.includes(role as Exclude<TUserRole, "USER">)) {
         delete query.status;
 
         if (role === UserRole.LIBRARIAN) query.librarianId = _id;
@@ -245,6 +245,17 @@ const deleteBookById = async (id: string) => {
     if (!book) {
       throw new NotFoundError("Book data not found!");
     }
+
+    const librarian = await User.findById(book.librarianId)
+      .select("books")
+      .session(session);
+
+    if (!librarian) {
+      throw new NotFoundError("Librarian not found!");
+    }
+
+    librarian.books.pull(id);
+    await librarian.save({ session });
 
     await Order.deleteMany({ bookId: id }).session(session);
